logger/pyproject.toml
[tool.poetry]
name = "logger"
version = "0.1.0"
description = ""
authors = ["Reinaldy Rafli <aldy505@tutanota.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = {extras = ["standard"], version = "0.16.0"}
fastapi = "0.70.1"
python-dotenv = "0.19.2"
influxdb-client = {extras = ["ciso"], version = "1.24.0"}
requests = "2.26.0"
pydantic = "1.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.12.2"
coverage = "^6.2"
pytest-race = "^0.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
